Separated NoteCard.kt from NotesScreen.kt
